+++ Tests/NablaMessagingCoreIntegrationTests/AuthenticatedEndpointsTests.swift
@@ -40,7 +40,7 @@ class AuthenticatedEndpointsTests: XCTestCase {
     func testCreateConversationFailsWhenNotAuthenticated() {
         _ = messagingClient.createConversation(
             title: nil,
-            providerIdToAssign: nil,
+            providerIds: nil,
             handler: assertAuthenticationErrorHandler()
         )
         waitForExpectations(timeout: 0.5)